LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/align-content-wrap-002-expected.txt


PASS .flex-horizontal, .flex-vertical 1 
FAIL .flex-horizontal, .flex-vertical 2 assert_equals: 
<div class="flex-horizontal" style="flex-wrap:wrap;">
    <div class="item-horizontal" data-expected-height="150"><div class="content1-horizontal"></div></div>
    <div class="item-horizontal" data-expected-height="150"><div class="content2-horizontal"></div></div>
    <div class="item-horizontal" data-expected-height="150"><div class="content3-horizontal"></div></div>
</div>
height expected 150 but got 80
PASS .flex-horizontal, .flex-vertical 3 
PASS .flex-horizontal, .flex-vertical 4 
FAIL .flex-horizontal, .flex-vertical 5 assert_equals: 
<div class="flex-vertical" style="flex-wrap:wrap;">
    <div class="item-vertical" data-expected-width="150"><div class="content1-vertical"></div></div>
    <div class="item-vertical" data-expected-width="150"><div class="content2-vertical"></div></div>
    <div class="item-vertical" data-expected-width="150"><div class="content3-vertical"></div></div>
</div>
width expected 150 but got 80
PASS .flex-horizontal, .flex-vertical 6 
Test for crbug.com/362848: Flex box word-wrap is not adhering to spec
